requirements.txt
# Dependencias principales del proyecto Digital Twin del Pórtico

# Computer Vision y Machine Learning
opencv-python==4.8.1.78
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0
Pillow==10.0.0

# Visualización y Gráficos
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
pyglet==2.0.9

# Interfaz Gráfica
# Nota: tkinter viene incluido con Python estándar
# Para sistemas que no lo incluyan por defecto:
# tkinter (incluido en Python estándar)
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3

# Comunicación y Networking
paho-mqtt==1.6.1
websockets==11.0.3
requests==2.31.0

# Manejo de Datos
pandas==2.0.3
h5py==3.9.0
openpyxl==3.1.2

# Configuración y Logging
PyYAML==6.0.1
coloredlogs==15.0.1
loguru==0.7.0

# Matemáticas y Física
sympy==1.12
numba==0.57.1

# Utilidades del sistema
tqdm==4.65.0
click==8.1.6
psutil==5.9.5
watchdog==3.0.0
pathlib2==2.3.7  # Para compatibilidad con versiones anteriores de Python

# Testing y Desarrollo
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Documentación
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Dependencias específicas para el proyecto
# Estas son necesarias para el funcionamiento completo del sistema
dataclasses-json==0.5.14  # Para serialización de dataclasses
typing-extensions==4.7.1  # Para compatibilidad de tipos

# Dependencias específicas del sistema (instalar según necesidad)
# Para Windows:
# pywin32==306

# Para Linux (instalar con apt/yum):
# python3-tk  # Para tkinter en sistemas Linux
# python3-dev
# build-essential

# Para macOS (instalar con brew):
# python-tk

# Notas de instalación:
# 1. Crear entorno virtual: python -m venv .venv
# 2. Activar entorno: 
#    - Windows: .venv\Scripts\activate
#    - Linux/Mac: source .venv/bin/activate
# 3. Instalar dependencias: pip install -r requirements.txt
# 4. Para desarrollo: pip install -r requirements.txt -e .
# 5. Si tkinter no está disponible en Linux: sudo apt-get install python3-tk
# 6. Para ejecutar el programa: python run_interface.py